Split out `variable_font_filename` into a reusable condition.
And use it on **com.google.fonts/check/canonical_filename**
(issue #3274)
